+++ src/test/api.test.ts
@@ -1,4 +1,4 @@
-import * as assert from 'assert';
+import assert from 'assert';
 import * as vscode from 'vscode';
 import { OmgLolApi } from '../api';
 import { AuthenticationManager } from '../authentication';